add class file
